Start of an acceptance testing environment
Use Kind (Kubernetes in Docker) to provide a genuine Kubernetes cluster
for testing purposes. We upload a copy of the manager into the cluster
using docker commands, which while brittle seems to work fine. We could
do with waiting for the manager statefulset to become ready, and to
support restarting the pod whenever we push in a new image.