Cache directory lookups, periodically refresh
The Google directory lookup APIs are quite expensive, and most users
will want to avoid hitting rate limits in normal operation. Without
caching, the number of API calls will scale linearly with the number of
groups used in all directory role bindings in the cluster: this is
likely to hit limits fast.

This change adds a cached directory role binding so that we cache
results on the group key. Membership changes are infrequent, so we
assume a default cache TTL as 5m. Using the cache means the number of
API requests made to the directory service will scale lineraly with the
number of **distinct** groups used across all directory role bindings in
the cluster, which could be a significant improvement.

Finally, we add a refresh parameter that causes the controller to
periodically reschedule reconciles, ensuring that changes to the group
membership in Google will be propagated to the role bindings in the
cluster.