Split managers by API Group
It's likely not practical to create a single manager that does
everything. Doing so would require the service account for that manager
to have broad permissions, and any failure in one manager might take
down the others.

While developing, you're also likely to want to be working on a single
API Group at a time, and want to run just that code change. Separate
managers would help here also.

This commit renames our previous 'manager' to be 'rbac-manager', which
will be responsible for managing all things under the rbac.lawrjone.xyz
group.