pkg/apis/workloads: Console (no codegen)
Create the definition of a Console, an instance of a console
environment. This is only the skeleton of what we'll have later, but is
designed to store the User field which we'll then prove can work with
mutating webhooks.

This commit does not update codegen, to make it clearer to reviewers
what has been done by a human and what is machine generated.