Skeleton workloads controller and webhooks
Create the new workloads manager, along with skeleton controller and
webhook code. This version of the manager uses automatic creation of
webhooks in the cluster, and also manages the service that will point at
the statefulset. Ideally we'd only automatically update the webhook
configuration but the controller-runtime doesn't allow this: we'll
likely change this implementation in future.