Implement expiration of console jobs
All consoles will have an expiration time, as we cannot support running
workloads in the cluster for an unconstrained amount of time: this will
both unnecessarily consume resources when interactive consoles are
detached, and hinder maintenance tasks which requires the eviction of
running pods, such as a node pool upgrade.

All consoles have an expiration time in their specification. Once the
console has been running for the amount of time defined by the timeout
the controller will delete the Job object. This deletion propagates to
the Pod which will be removed too, according to its termination grace
period.