Remove Recreate operation from reconcile.CreateOrUpdate
The way this function currently works, it's not actually possible to
perform with a Recreate operation: we look up the resource by name, so
the name will always be the same.

We may need this in the future when we have resources that can change
name, but for now it is not needed.