Add a console Runner package
This is a wrapper around logic required to create and attach to
consoles, to be called from external clients.

The code here enables creation of a very basic console, and then waiting
for it to become ready to attach to.

The interfaces are not necessarily stable at this point and subject to change!