reconcile.ResolveAndReconcile
Introduce a new function called ResolveAndReconcile that should help us
make our reconciliation loops more consistent, while reducing
boilerplate.

Almost every reconcile loop starts by fetching the object that triggered
it, and handling errors if that object is not found or worse.
Additionally we setup the logging to direct events to the Kubernetes
event trail which we want to ensure is in place for every resource we
reconcile.