Fix negative RequeueAfter value
We cannot pass a negative value for `RequeueAfter`, as if we do this
then it is not handled correctly by the controller-runtime
reconciliation code, and we don't actually get a reconcile.

We can get into a situation where reconciliation is overdue if the
reconcile loop is called when the console is running, but after the
expiry time is hit.