Watch Job events and reconcile for an owner
Watch for Job events and enqueue a reconcile.Request for the Console
owner. When a Job is Completed, Failed or ActiveDeadlineSeconds has
been reached, a job event is fired following by a reconcile for the owner.

Add jobSpec defaults to avoid multiple update jobs in the reconcile
loop.

Update jobDiff to cover all mutable fields in the jobSpec