Always set a console pod's RestartPolicy to Never
In order to more easily reason about the lifetime of our console tasks,
we never want a container that fails with a non-zero exit code to [be
restarted][1].

The only other supported value for `RestartPolicy` in Jobs is
`OnFailure`, which we won't want to use, so setting it to `Never` in the
controller is a sensible step.

This saves us from having to specify a RestartPolicy in each template
that we define.

[1]: https://kubernetes.io/docs/concepts/workloads/controllers/jobs-run-to-completion/#handling-pod-and-container-failures