WaitUntilReady waits for RoleBinding
The Console controller provisions a DirectoryRoleBinding to bind the
console user (and any additional attach subjects) to a role that grants
them permission to attach to the console pod. The DirectoryRoleBinding
controller then provisions a RoleBinding, and then updates it with the
subject list.

Until this RoleBinding exists, and contains the user in its subject
list, the user will not be able to use their console.