theatre-envconsul, install & exec
Create a binary that can be used by a mutating webhook to instrument
Hashicorp's envconsul.

The install command copies the theatre-envconsul and envconsul binaries
from the root filesystem (where they are expected to be installed by our
docker image) to an install path. We intend to run this as an init
container.

The exec command then authenticates with Vault, exchanging a kubernetes
service account token (detected from our environment) for a Vault token.
We then parse user configuration files which contain an environment
variable map, and produce an envconsul file that will resolve vault
secrets for those environment variables that have a `vault:` prefix. We
finally os.Exec to the envconsul command, which then takes over.

We expect the exec command to be prepended to pod container commands.

Co-Authored-By: Dyson Simmons <dyson@gocardless.com>