Config for vault manager
This adds the deployment config for the vault-manager and related
resources. The vault-manager is deployed as a StatefulSet as we only
want a single replica of it.

We use Kustomize vars to grab the deployment image and pass it to
vault-manager, which runs the mutating webhook that injects the init
container running this same image.

Co-authored-by: Dyson Simmons <dyson@gocardless.com>