Envconsul webhook
Adds a mutating webhook that receives requests on pod creation. It
checks for the presence of the envconsul-injector annotation on a
pod and if it is found it will mutate the pod by injecting in an init
container, adding a volume and mount, and modifying the command.
Containers without this annotation are returned unmutated.

The init container is used to copy the theatre-envconsul and envconsul
binaries into a volume that is shared with the configured containers in
the annotation.

The command of the configured containers is then modified to run our
theatre-envconsul binary. This binary authenticates from vault, reads
the configuration file if specified, stamps out an envconsul config, and
then syscall.Exec's the envconsul binary. This binary runs as PID 1,
retrieves the specified secrets from vault, sets them to environment
variables, and spawns the original container's command as a child
process.

Co-authored-by: Dyson Simmons <dysonsimmons@gmail.com>
Co-authored-by: Lawrence Jones <lawrjone@gmail.com>
Co-authored-by: Ben Wheatley <contact@benwh.com>