Acceptance tests for theatre-envconsul
Provision a Vault instance inside the Kind cluster. We can then use this
instance to validate the behaviour of theatre-envconsul by running a pod
with a specific service account that outputs its environment.

We don't yet test the configuration file for theatre-envconsul, nor the
mutating webhook. It would make sense to add these later.

To get this working, I needed to amend our acceptance environment to
boot a cluster using a custom kind-e2e.yaml config. While there, I've
changed the dockerfile we use to generate a test image to use what we
will for production, as this docker image contains things other than
just our binaries now (envconsul). This comes at the cost of being
slower to build, but is the correct way to do things.

Configuring Vault with appropriate policies is done using
theatre-envconsul itself, as the acceptance tests aren't guaranteed to
be run within the cluster where the service is accessible. It might make
sense for us to vendor a terraform file that can provide this
configuration in future, something that matches what we use internally
at GC.