Accept command as args in theatre-envconsul
This makes it such that the command to be executed by theatre-envconsul is taken in
as command line args rather than via a --command flag. This should provide a nicer
interface as that's the command we want theatre-envconsul to exec.

We also need to update the mutating webhook to reflect this change.