Update production image tag
We moved the binaries in the docker image to be under /usr/local/bin/
and updated the kustomise to point to these locations. We did't however
update the production image tag. This meant we were deploying a old
image where the binaries were root of the file system but the kustomize
was expecting them in the new location.