Increase vault-manager replicas
We need to run more replicas of the vault-manager to ensure high availability,
as we need the mutating webhook to be always available - new pods will not be
created otherwise. If the manager ever needs to manage custom resources in the
future, we need to implement leader election in vault-manager to prevent races.