Use large resource_class in CI for acceptance
We require more CPU to spin up all off the kubernetes resources in the
acceptance test step. There are only two machine types in Circle CI so
setting this to large. We could alternatively patch the resources in the
acceptance overlay to keep them under the default machine resource
limits.