Acceptance test runners, with optional prepare steps
For acceptance tests, it is good to be able to prepare some additional
resources prior to running the tests themselves. To do this, refactor
our existing acceptance test suites to use a Runner interface, where
each suite can provide optional prepare commands.

This allows an operator to run cmd/acceptance prepare to re-converge
their cluster, improving the local developer experience. Using this
pattern enables a refactor of lots of test-only functionality from the
theatre-envconsul binary, which is a win for readability.