Hardcode vault-manager theatre image tag
We originally used kustomize vars to run vault-manager with the version that's
used in the production overlay.

Unfortunately, kustomize v1.0.11 that we run in anu does not support fieldRef
indexing (ie. spec.template.spec.containers[0].image) in vars. Therefore, the
image tag for vault-manager is hardcoded for now . We need to bump this manually
to keep it in sync with latest deployed version of theatre.

TODO: tidy this up using jsonnet