theatre-envconsul base64-exec
Add a hidden command that can be used to ensure we preserve shellword
splitting. The readme covers how we do this, but in picture form:

```
$ theatre-envconsul exec -- <command>
  -> envconsul --config rendered.json (with base64 command)
  -> theatre-envconsul base64-exec <base64-encoded-command>
  -> <command>
```

theatre-envconsul now renders an envconsul config that will call
theatre-envconsul base64-exec, with the original command arguments
base64-encoded, rather than the command directly. We lean on theatre to
decode the arguments and exec the command as originally shell-split.