Remove check for presence of secrets
theatre-envconsul assumes that your app needs secrets when you opt-in by
annotating your service. when it doesn't find the vault: prefix, it errors
as envconsul hangs when provided no secrets. we have now changed that
assumption as gc-app now adds that annotation to every service, causing apps
that don't use the vault: prefix (currently everything) to crash loop.

with this change in our fork of envconsul, we no longer have to worry about
envconsul hanging when no secrets are provided.