Use request namespace and name if it exists
A pod sent to our envconsul mutating webhook doesn't always have the
namespace or name in the pod spec, but the request may. We can always
rely on the request namespace being the accurate of the two so we should
always use it. Unfortunately pods usually don't contain a name, and the
request most likely doesn't either, but if it does we should also
prefer it over what is set in the pod.

As pod requests a most likely to be made from a deployment, which
doesn't allow the setting of a pod name in the pod template, we can
expect in most cases for the pod name to be blank. This makes it hard to
log useful context, such as the pod name, in the mutating webhook.