Verify non-root user can access secrets
Add a test that confirms a non-root docker container can access Vault
using the projected service account token. This is verifying that the
fsGroup assignment is having the effect we want it to, and providing the
non-root user with a supplementary group that permits reading the token
on the mounted volume.