Add theatre-consoles and initial create command
This change introduces a theatre-consoles kingpin application.
The application has a single command, ``create`` which will create a
console instance, ready to attach to.

This is part of a series of changes to provide a command line to
work with consoles in theatre. The management logic currently in the
consoles application will likely move to be closer to the runner in
the near future.