Add authorisation support to console CRDs
Adds fields to the ConsoleTemplate CRD that facilitates implementing
peer-reviewed consoles.

`authorisationRules` provides a list of rules that commands can be
matched against and if a match is found, the console requires
authorisation to run from a defined number of subjects. It is intended
that the rules will be matched from top to bottom and the first match is
used.

`defaultAuthorisationRule` provides a default rule to match on if no
other rules match.

Both these fields are optional allowing for backward compatibility.

An additional CRD has also been added to facilitate the authorisation
process. The `ConsoleAuthorisation` kind provides a way to monitor
approvals provided to a console that matches a rule, or the default
rule.

The actual implementation of the peer-reviewed console feature is to
follow.