acceptance: Support kind v0.7.0
In the latest release of `kind`, the `kind get kubeconfig-path`
subcommand has been removed. See PR #1143 in the kind repo for further
info.

This has been replaced, to some extent, with `kind export kubeconfig`,
which sets the context to the kind cluster, and `kind get kubeconfig`:
but this dumps the whole kubeconfig to stdout, not just the path.

It is still possible to use a separate `KUBECONFIG` file, by having this
variable exported when creating the kind cluster, but there should be no
issue with just having the context added to the main `KUBECONFIG` file,
and this seems to be what the project is suggesting.

Therefore just change the invocation of commands, when we shell out, to
use an explicit `--context`, rather than overriding `KUBECONFIG`.