Add namespace and context flags to the cli entrypoint
This change introduces a ``namespace`` and a ``context`` flag to allow
the user to specify the relevant namespace and/or kubernetes context
for the invocation of the command.
This solves the issue of running a console when multiple templates
match the selectors across namespaces, in addition to making it more
explicit and easier to include in a script.