Reorder Dockerfile for quicker builds
When developing this application we are constantly invalidating the
first layer of the container, by changing the source files. This causes
us to have to rebuild `envconsul`, which takes up to a minute.

Fix this by building `envconsul` first, so that we'll keep this cached
and only need to repeat this step if changing the envconsul version.