Clean up console integration test
While adding new specs to this suite it became clear that it's not laid
out ideally.

The primary change here is to split out configuration and creation of
resources, as per the style [suggested in the Ginkgo docs][0].

What this means is that it's now possible to override parts of the
console template or console within each `Describe` block, leading to
much cleaner code as the specs themselves now no longer need to create
resources before asserting against them.

Due to the introduction of some more nesting, it may be easier to review
this diff without whitespace changes.

[0]: https://onsi.github.io/ginkgo/#separating-creation-and-configuration-justbeforeeach