Don't require update to authorisers in webhook
We discovered, while writing reconciliation code, that if you were to
perform an update against this resource but not add any authorisers,
then the update would be rejected.

This is unintuitive, it should be possible to perform operations such
as adding annotations or altering non-core fields, without any issue.

Change the validation code to only throw an error if there's greater
than 1 addition, as opposed to 0 or >1 additions.