Add reconciliation of authorisation RBAC objects
Stamp out the RBAC objects that allow console authorisers the ability to
update the `consoleauthorisation` object with their approval.

Which subjects are granted access to update this is dependent on the
authorisation rules in the template and the command that is matched to
a rule; so also add the helper function to determine this.