Fix console job name truncate length
Jobs names are restricted to 63 characters as they are added as a label
to the pods they create (and labels have a 63 character length
restriction).

We incorrectly limited job names thinking that the pods they created had
a name limit of 63 characters. This fix gets us a few characters back in
the job name and ensures our labels are the correct length.