Prevent console jobs recreating on delete
Only create jobs on console creation and mark a console as destroyed if
a console job has been deleted from underneath the console.

Before this fix, deleting a job (even if the console had completed and
was waiting for GC) would result in a new job being created for the
console and therefore a new console pod executing.

The recreation of the job is unexpected behaviour and this change
ensures this double running cannot happen.