Add golangci-lint to CI
This change introduces the GolangCI linter into the CI, alongside
some sensible defaults for enabled linters.
This should help ensure that we are producing high quality code,
whilst raising potential issues

It is important to note that these configurations are not set in stone,
and are likely to change to suit the workflow of the project, enabling or
disabling behaviour as seen fit.