Move commands from cmd to pkg/workloads/consoles/runner
This change moves the commands previously defined in the
``cmd/theatre-consoles`` package to ``pkg/workloads/console/runner``.
These functions are helpers to interact with consoles easily, and with
this change ``cmd/theatre-consoles`` becomes a very thin layer mapping
flags to the underlying functionality.

This enables these commands to be embedded easily inside of
other applications, using a standard interface.