Remove owner from console authorisation
Remove the owner from the console authorisation and instead look up the
console via the consoleRef to get the console user.

It was thought that having the owner of the console as a field in the
console authorisation would simplify the implementation and while it did to
a small degree it made the console authorisations difficult to test.

Kind only previsions a single user when creating the cluster and we are
unable to provision more. This prevents us from being able to submit an
authorisation as another user - users can only add themselves to the
authorisations and a user cannot authorise their own console.

We do support service accounts as authorisation subjects but
unfortunately impersonation in the go client doesn't support service
accounts and so we can't impersonate one when submitting the
authorisation.

As console authorisations require updating from end users, the owner
fields immutability was enforced on the console authorisations
validating webhook. This left us with no way to modify it.

Removing the owner field and looking up the user from the consoleRef
allows us to test the authorisation flow and is arguably an improvement,
as now we don't have state split over two objects that could get out of
sync.

As our implementation of consoles requires that users don't have
permission to update consoles, we can trust that the user in the console
is correct. And as we don't lock down consoles in the kind test cluster,
we can update the consoles users to another user before performing a
validation.