Add console authorisation phases to reconcile
Adds the concept of authorisation to the console life-cycle.

If a console template has no rules, then a console is automatically
authorised and it functions the same as it did before this change.

If a console template has rules, it now waits for authorisation with a
`Pending Authorisation` phase. On each update the to the console
authorisation the console will reconcile again and it will check if the
number of authorisations matches the required number of authorisations
specified in the rule the command matches (or the default rule). When
the number has been met, the console console is allowed to continue to
`Pending` as normal.