Add pending approval TTL and improve GC
With the introduction of authorisations consoles that need
authorisations enter a Pending Authorisation state. If the console is
never authorised, we need to ensure we clean them up and don't just
leave them hanging forever.

We introduce a new field on the console template (and therefore
template) to set the TTL for a console to reach the Running phase. If a
console doesn't reach this phase before the TTL (including if it is
still in the Pending Authorisation phase) then we garbage collect it.