Add note about ns deletion in integration tests
I was under the impression that the per-spec namespaces were removed
after each test run, but this is not the case as per the comment below
the added line.

Add a log message to remind the reader of integration test logs that
this isn't happening. This will hopefully make it clearer as to why
there are log entries from different integration specs' reconciliations
interleaved with each other.