Fix nil pointer dereference in console logging
This has been breaking our integration tests intermittently for some
time.

The issue was in the following conditional:
```go
if newStatus.CompletionTime != r.console.Status.CompletionTime && !statusCtx.Job.Status.StartTime.IsZero() {
...
```

In the first part of the expression we were comparing two pointers
(`CompletionTime`) for inequality. As far as I can tell this is *always*
going to evaluate to true, because they're different objects, so was
potentially wrong anyway.

We then get to the second part of the expression and hit a nil pointer
dereference because the `Job` field is not populated on `statusCtx`.
This is actually valid: It will happen when the console is in the
`Destroyed` phase (the job has been deleted).

Fix this by changing the conditional: Inspect the new status' phase and
only log this event if it's equal to `Stopped`.

This leaves us open to potentially not logging this `ConsoleEnded` event
if the console has effectively transitioned through multiple phases
between a reconcile loop; for example from `Pending` to `Running` to
`Stopped`. I'm not sure whether this is actually possible in a real
deployment, but add a new block to capture this, in case it is.