Basic support for multiple containers in console
Console templates by design only allow a single container to be defined.
This restriction then relied on the default behaviour of kubernetes - if
there is only a single container, don't require specifying the container
in actions such as attaching or displaying logs.

Relying on the default behaviour breaks when containers are injected into the
console pod via webhooks. In this case kubernetes requires you to also
provide the name of the container that you want to interact with.

This change gets the first container that has a TTY (which should only
be the single container defined in the console template) and passes it
through to attach.

Ideally we should support multiple containers in consoles and allow the
specifying of a container to attach to from the command line as well as
being able to set a default in the console template. This change however
is the minimum to support multiple containers without changing current
behaviour.