Remove generated clients
Kubebuilder has made the decision to move away from generated, typed
clients, in preference of using a client that is backed by a scheme
store. (https://github.com/kubernetes-sigs/kubebuilder/issues/1152#issuecomment-584774562)

This will need to change in quite a few places to use the new client
instead, as the signatures are different. Once completed this leaves
just the console webhooks and the tests to finish.