Move to new client in consoles/runner
This change moves to using the controller-gen client in the runner
library. This is one option being evaluated as the kubebuilder project
has discounted the idea of generatingclients in projects using it's
structure.

An alternative approach may be to extend the client code to support
watching, which could potenitally be included upstream, or we could
attempt to work into a structure that facilitates creation of a
client, with the knowledge that supporting this goes against the
grain, and may cause us issues in a future update.

This commit moves the runner tests to disqualify them as go code until
the new testing strategy is in place.