WIP: Add back acceptance tests
This is a dump of all my changes just to have a backup somewhere in case
I lose them! This branch will need to be redone into a workable PR in any
case so committing this horrible work in progress for now.

Also many other small changes here and there where I have come accross
things to fix.