Add workloads integration tests
Adds a simplified version of the existing console integration tests.

It doesn't test the console webhooks as webhook integration in envtest
isn't as simple as the crds yet. This was previously ok because we had a
manager that would create the webhook configuration objects and handle
self serve certificates for us which is no longer the case. These
webhooks are tested both in unit tests and acceptance tests in any case
and the tests here for them didn't provide a lot of coverage.

Also no longer uses the advanced features that were in our
pkg/integration pkg. Without webhooks this pkg was half as useful, and
with all the changes between v1 and v2, the work to resolve the rest of
the features at the moment seemed to be of little benifet. We could
perform the same tests without the extra complexity.