Progress on fixing acceptance tests
The cceptance tests now progress further, though not yet to
completion. The pod created by the job for the consoles acceptance
test no longer appears to run, which causes the new failure point.

The fix for the previous failure point was the status being a
subresource, this meant that the status was never updated with the
rest of the object. However, in the future, it may make the reconcile
loop easier to work with if we refactor the code to make it more
explicit where we want to perform actions.