Remove all non-rbac resources
The upgrade to kubebuilder v2, along with the upgrade of
controller-runtime and controller-gen is extremely invasive and cannot
easily be done on top of our existing project.

In order to make the initial change to the new structure, we have chosen
to isolate the RBAC controller. As we can't leave half of the old v1
project around, we need to remove all of the non-rbac related resources.

This should allow us to create a nice diff for rbac in the next commit
where we move rbac to v2.

This is a best effort approach to moving to v2 with minimal pain. We
will add these deleted files back for each resource as we add them to
maintain the diffs.

Co-authored-by: Jack Stephenson <jackatbancast@users.noreply.github.com>