Migrate rbac controller to kubebuilder v2
As rbac is one of the simplest components, as well as being a dependency
of other components in theatre, we have chosen to migrate it to v2
first.

This commit also includes the surrounding changes to things such as
docker, circleci, Makefile, etc, which make up the project and are
necessary.

Co-authored-by: Jack Stephenson <jackatbancast@users.noreply.github.com>