v2: Disable make targets and ci jobs
Disables a bunch of make targets & circle ci steps so no code is built,
tested, or generated.

As we have deleted vendor we are in a broken state and we only want to
add these back as we add v2 features (ignoring the as yet to be migrated
v1 files). This way as we move over each component we can move over
fully working tests, code generation, and ci steps.