v2: rbac - add go vet to makefile and ci
Adds a target to run go vet in Makefile and also adds a vet step to
CircleCI. Most of the errors go vet detects are found at compile time,
but as per the docs, it can find some the compiler can't. Best to ensure
we run it in CI and that it passes.