Use the provided message instaid of the error string
This change changes the behaviour of the recorder to use a provided
error message instead of the string representing the error. In most
cases this is currently the same as the message is passed in as
err.Error(), though this is a choice that should be made in the code,
not by the recording library.