v2 vault: Changed resources back to statefulset
Kubebuilder uses by default deployment with multiple replicas and leader
election. We never run theatre in that setup, so they're some unknows
and risk related to that changed.

This changed amin to reduce the blast radius and ensure that we only
focus in the minimun changes that we need to update to kubebuilder v2.

This commit can revert in the future to switch to deployments + leader
election in a future changed.