Fixup: add workloads ctlr tests to Makefile
The `controller/workloads/console/integration` tests were left out of
the `make test` target. They were correctly added to ci.

We will be moving back to globbing for all of these make targets/ci jobs and so it would have
been picked up again when that happens but for now lets add it back to
make reviewing easier.