Introduce noninteractive consoles
The assumption on the initial implementation of consoles was that making
consoles universally interactive, with STDIN and TTY enabled on their
containers, would cover all use cases. But it turns out that the wrapper
binary that Tekton uses to co-ordinate steps in a task does not cope
well with TTY: true. So we introduce a noninteractive form of console
attachment here that will play nicely with Tekton and may be a more
suitable target for other automation that uses the console resource.

It is now also the case that the attach and create subcommands exit with
an error if the console process fails; previously you would get a 0.