Delete theatre-envconsul acceptance test
The acceptance test for theatre-envconsul and the vault-manager are
copies of each other. The correct tests and the ones that are actually
configured to be used are the tests in vault-manager. These test also
cover the happy path of theatre-envconsul.

The unused and duplicated acceptance tests in theatre-envconsul can be
deleted.