Remove the lease option from theatre-envconsul
Remove the lease feature flag from theatre-envconsul and remove the code
associated with it.

Allowing envconsul to spawn and own our process as it's child has caused
us a number of issues as in many circumstances envconsul resorts to
restarting the child process.

We don't want this behaviour and we instead want the child process to be
spawned and left alone. As such we don't want to leave the lease option
in as it is likely to lead anyone else who uses this down a similar path
of issues.