Fix splitting env vars in theatre-envconsul env
os.Environ() returns the processes environment variables as a slice of
strings. The name of the environment variable and the value of it are
joined with an "=" character. For example "MY_ENV_VAR=my-env-var-value"

Inside of `theatre-envconsul env` we ranged over this slice splitting on
the "=" character to separate the name from the value. Incorrectly the
splitting wasn't limited to the first occurrence of the "=" character and
so environment variables that contained an "=" character would be
truncated. For example "MY_ENV_VAR=my=env=var=value" would become
"MY_ENV_VAR=my".

Many environment variables contain an "=" character such as passwords,
tokens, and base64 encoded data. This change limits the split to the
first occurrence preserving the value correctly.