theatre-envconsul: Log envconsul output upon failure
Previously if `envconsul` returned a non-zero exit code, then we'd only
see the exit code number in the error output.

After this change we include the `stderr` output from the `envconsul`
binary, which helps to debug the root cause without having to
manually re-run `envconsul` with the config file as a parameter.