Add check for pod health to runner attach
This change adds a check for pod health in the condition that
attaching to the pod fails.

This issue can occur in cases where failure to attach to the console
is due to issues with the pod, which manifest as RBAC issues to
kubernetes.
The addition of the health check should make this more apparent when
issues occur that are unrelated to the process of attaching to the console.